Add unit test for issuer-signed cbor data
