+++ Tests/WalletStorageTests/WalletStorageTests.swift
@@ -1,0 +1,23 @@
+import XCTest
+import Foundation
+@testable import WalletStorage
+import MdocDataModel18013
+
+final class eudi_lib_ios_wallet_storageTests: XCTestCase {
+	
+		
+    func testIssueRequestPublicKey() throws {
+       let ir = try IssueRequest()
+       let pem = try ir.getPublicKeyPEM()
+       XCTAssert(Bool(pem.count > 0))
+       print(pem)
+    }
+	
+	func testLoadPidIssuerSigned() {
+		let base64str = String(data: Data(name: "eu_pid_base64", ext: "txt", from: Bundle.module)!, encoding: .utf8)!
+		let pidIssueredSigned = Data(base64Encoded: base64str)!
+		print(pidIssueredSigned.count)
+	}
+	
+		
+}
